feat: refactor code to improve structure and readability
